Made second argument optional and -m the default
The second argument for install is now optional and defaults to the folder name of the source (or the repository name for GitHub sources).

The -o option is now an optional second argument for the use command.

The use command now defaults to merging the new template into the OUT directory. -f was renamed to --hard and forces the deletion and creation of the out directory (if it exists).