Configuration management refactored
The config file can now be passed to boil by the environment (BOIL_CONFIG), by option (—config) or be loaded from the default config file. The temaplte directory cann additionally be set with —tpldir.

The config is also no longer stored in a global object, but in the click context object.

A new pass_tpldir decorator passes the tpldir directly to a command that does not need the full context object.